+++ game/models.py
@@ -45,7 +45,7 @@ class Episode(models.Model):
 
     name = models.CharField(max_length=200)
     next_episode = models.ForeignKey(
-        "self", null=True, default=None, on_delete=models.SET_NULL
+        "self", null=True, blank=True, default=None, on_delete=models.SET_NULL
     )
     in_development = models.BooleanField(default=False)
 
@@ -59,6 +59,19 @@ class Episode(models.Model):
     r_pythonEnabled = models.BooleanField(default=False)
     r_trafficLights = models.BooleanField(default=False)
     r_cows = models.BooleanField(default=False)
+
+    lesson_plan_link = models.CharField(
+        max_length=500, null=True, blank=True, default=None
+    )
+    slides_link = models.CharField(
+        max_length=500, null=True, blank=True, default=None
+    )
+    worksheet_link = models.CharField(
+        max_length=500, null=True, blank=True, default=None
+    )
+    video_link = models.CharField(
+        max_length=500, null=True, blank=True, default=None
+    )
 
     @property
     def first_level(self):
@@ -86,12 +99,23 @@ class Episode(models.Model):
             7: "medium-hard",
             8: "medium-hard",
             9: "brainteasers",
-            10: "split-python",
-            11: "split-python",
-            12: "python",
-            13: "python",
-            14: "python",
-            15: "python",
+            10: "early-python",
+            11: "early-python",
+            12: "late-python",
+            13: "late-python",
+            14: "late-python",
+            15: "late-python",
+            16: "early-python",
+            17: "early-python",
+            18: "early-python",
+            19: "early-python",
+            20: "late-python",
+            21: "late-python",
+            22: "late-python",
+            23: "late-python",
+            24: "late-python",
+            25: "late-python",
+            26: "late-python",
         }
 
         return difficulty_map.get(self.id, "easy")
@@ -140,6 +164,7 @@ class Level(models.Model):
     next_level = models.ForeignKey(
         "self",
         null=True,
+        blank=True,
         default=None,
         on_delete=models.SET_NULL,
         related_name="prev_level",